统一tools中env相关的接口到env_utility.py文件，并同步调整ci
