+++ .github/workflows/compile_bsp_with_drivers.yml
@@ -52,7 +52,6 @@ jobs:
           git config --global http.postBuffer 524288000
           git remote -v
           git fetch origin
-          python -c "import tools.menukconfig; tools.menukconfig.touch_env()"
 
       - name: Install Arm ToolChains
         if: ${{ success() }}
@@ -67,5 +66,5 @@ jobs:
         if: ${{ success() }}
         shell: bash
         run: |
-          source ~/.env/env.sh
-          #python tools/ci/compile_bsp_with_drivers.py
+          # source ~/.env/env.sh
+          # python tools/ci/compile_bsp_with_drivers.py
